feat(issues): Prevent loading grouping info
- on streamline ui, no longer loads grouping info when collapsed. We can't do this on the old UI because we always display a "Grouped By".
- Lazy loads grouping info components
